Uploading directory of all project files
bitonicSort.java is basic, uses arrays to sort any array given. - Doesn't work unless it's a power of two and can't pad because it's an array
bitonicSort2.java is identical to bitonicSort.java but showcases what happens when you try to sort an array that isn't a power of 2
finalBitonicSort.java uses an ArrayList<Integer> to sort - Allows padding to power of 2 to sort successfully. Then cleans up output and deletes all 0's